+++ DSCResources/MSFT_xSQLServerScript/MSFT_xSQLServerScript.psm1
@@ -31,6 +31,10 @@ Import-Module -Name (Join-Path -Path (Split-Path -Path (Split-Path -Path $script
         Use a Windows PowerShell array to specify multiple variables and their values. For more information how to use this,
         please go to the help documentation for [Invoke-Sqlcmd](https://technet.microsoft.com/en-us/library/mt683370.aspx).
 
+    .PARAMETER QueryTimeout
+        Specifies, as an integer, the number of seconds after which the T-SQL script execution will time out.
+        In some SQL Server versions there is a bug in Invoke-Sqlcmd where the normal default value 0 (no timeout) is not respected and the default value is incorrectly set to 30 seconds.
+
     .OUTPUTS
         Hash table containing key 'GetResult' which holds the value of the result from the SQL script that was ran from the parameter 'GetFilePath'.
 #>
@@ -60,12 +64,16 @@ function Get-TargetResource
         [System.Management.Automation.Credential()]
         $Credential,
 
+        [Parameter(Mandatory = $false)]
+        [System.UInt32]
+        $QueryTimeout,
+  
         [System.String[]]
         $Variable
     )
 
     $result = Invoke-SqlScript -ServerInstance $ServerInstance -SqlScriptPath $GetFilePath `
-                -Credential $Credential -Variable $Variable -ErrorAction Stop
+                -Credential $Credential -Variable $Variable -QueryTimeout $QueryTimeout -ErrorAction Stop
 
     $getResult = Out-String -InputObject $result
 
@@ -75,6 +83,7 @@ function Get-TargetResource
         GetFilePath = [System.String] $GetFilePath
         TestFilePath = [System.String] $TestFilePath
         Credential = [System.Object] $Credential
+        QueryTimeout = [System.UInt32] $QueryTimeout
         Variable = [System.String[]] $Variable
         GetResult = [System.String[]] $getresult
     }
@@ -107,6 +116,10 @@ function Get-TargetResource
         to the built-in parameter `PsDscRunAsCredential`. If both parameters `Credential` and `PsDscRunAsCredential` are not assigned,
         then SYSTEM account will be used to authenticate using Windows Authentication.
 
+    .PARAMETER QueryTimeout
+        Specifies, as an integer, the number of seconds after which the T-SQL script execution will time out.
+        In some SQL Server versions there is a bug in Invoke-Sqlcmd where the normal default value 0 (no timeout) is not respected and the default value is incorrectly set to 30 seconds.
+
     .PARAMETER Variable
         Specifies, as a string array, a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
         Use a Windows PowerShell array to specify multiple variables and their values. For more information how to use this,
@@ -137,12 +150,16 @@ function Set-TargetResource
         [System.Management.Automation.Credential()]
         $Credential,
 
+        [Parameter(Mandatory = $false)]
+        [System.UInt32]
+        $QueryTimeout,
+  
         [System.String[]]
         $Variable
     )
 
     Invoke-SqlScript -ServerInstance $ServerInstance -SqlScriptPath $SetFilePath `
-                -Credential $Credential -Variable $Variable -ErrorAction Stop
+                -Credential $Credential -Variable $Variable -QueryTimeout $QueryTimeout -ErrorAction Stop
 }
 
 <#
@@ -169,6 +186,10 @@ function Set-TargetResource
         The credentials to authenticate with, using SQL Authentication. To authenticate using Windows Authentication, assign the credentials
         to the built-in parameter `PsDscRunAsCredential`. If both parameters `Credential` and `PsDscRunAsCredential` are not assigned,
         then SYSTEM account will be used to authenticate using Windows Authentication.
+
+    .PARAMETER QueryTimeout
+        Specifies, as an integer, the number of seconds after which the T-SQL script execution will time out.
+        In some SQL Server versions there is a bug in Invoke-Sqlcmd where the normal default value 0 (no timeout) is not respected and the default value is incorrectly set to 30 seconds.
 
     .PARAMETER Variable
         Specifies, as a string array, a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
@@ -202,6 +223,10 @@ function Test-TargetResource
         [System.Management.Automation.Credential()]
         $Credential,
 
+        [Parameter(Mandatory = $false)]
+        [System.UInt32]
+        $QueryTimeout,
+
         [System.String[]]
         $Variable
     )
@@ -209,7 +234,7 @@ function Test-TargetResource
     try
     {
         $result = Invoke-SqlScript -ServerInstance $ServerInstance -SqlScriptPath $TestFilePath `
-                -Credential $Credential -Variable $Variable -ErrorAction Stop
+                -Credential $Credential -Variable $Variable -QueryTimeout $QueryTimeout -ErrorAction Stop
 
         if($null -eq $result)
         {
@@ -243,6 +268,10 @@ function Test-TargetResource
         to the built-in parameter 'PsDscRunAsCredential'. If both parameters 'Credential' and 'PsDscRunAsCredential' are not assigned, then
         the SYSTEM account will be used to authenticate using Windows Authentication.
 
+    .PARAMETER QueryTimeout
+        Specifies, as an integer, the number of seconds after which the T-SQL script execution will time out.
+        In some SQL Server versions there is a bug in Invoke-Sqlcmd where the normal default value 0 (no timeout) is not respected and the default value is incorrectly set to 30 seconds.
+
     .PARAMETER Variable
         Creates a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
 #>
@@ -262,6 +291,10 @@ function Invoke-SqlScript
         [System.Management.Automation.Credential()]
         $Credential,
 
+        [Parameter(Mandatory = $false)]
+        [System.UInt32]
+        $QueryTimeout, 
+  
         [System.String[]]
         $Variable
     )